+++ apps/cli/src/commands/device/listApps.ts
@@ -1,16 +1,14 @@
 import { from, tap } from "rxjs";
 import { withDevice } from "@ledgerhq/live-common/hw/deviceAccess";
 import listApps from "@ledgerhq/live-common/hw/listApps";
-import { deviceOpt } from "../../scan";
+import { DeviceCommonOpts, deviceOpt } from "../../scan";
+
+export type ListAppsJobOpts = DeviceCommonOpts;
 
 export default {
   description: "list all installed apps on the device",
   args: [deviceOpt],
-  job: ({
-    device,
-  }: Partial<{
-    device: string;
-  }>) =>
+  job: ({ device }: ListAppsJobOpts) =>
     withDevice(device || "")(t =>
       from(listApps(t)).pipe(
         tap(res => {