+++ apps/ledger-live-desktop/tests/page/abstractClasses.ts
@@ -1,4 +1,5 @@
 import { Page } from "@playwright/test";
+import { step } from "tests/misc/reporters/step";
 
 export abstract class PageHolder {
   constructor(protected page: Page) {}
@@ -26,7 +27,8 @@ export abstract class Component extends PageHolder {
     return await this.page.waitForLoadState("load");
   }
 
-  async waitForPageNetworkidleState() {
+  @step("Wait for network calls to be completed")
+  async waitForPageNetworkIdleState() {
     return await this.page.waitForLoadState("networkidle");
   }
 }