+++ apps/ledger-live-desktop/tests/specs/speculos/manage.ledgersync.spec.ts
@@ -27,10 +27,7 @@ test.describe(`[${app.name}] Sync Accounts`, () => {
       await app.ledgerSync.expectSyncAccountsButtonExist();
 
       await app.ledgerSync.syncAccounts();
-      await app.speculos.clickNextUntilText("Make sure");
-      await app.speculos.confirmOperationOnDevice("Connect with");
-      await app.speculos.clickNextUntilText("Your crypto accounts");
-      await app.speculos.confirmOperationOnDevice("Turn on sync?");
+      await app.speculos.activateLedgerSync();
       await app.ledgerSync.expectSynchronizationSuccess();
       await app.ledgerSync.closeLedgerSync();