+++ apps/ledger-live-desktop/tests/specs/speculos/swap.spec.ts
@@ -319,8 +319,10 @@ async function performSwapUntilQuoteSelectionStep(
   electronApp: ElectronApplication,
   swap: Swap,
 ) {
+  await app.layout.goToAccounts();
+  await app.accounts.navigateToAccountByName(swap.accountToDebit.accountName);
   await app.layout.goToSwap();
-  await app.swap.waitForPageNetworkidleState();
+  await app.swap.waitForPageNetworkIdleState();
   await app.swap.selectAssetFrom(electronApp, swap.accountToDebit);
   await app.swapDrawer.selectAccountByName(swap.accountToDebit);
   await app.swap.selectAssetTo(electronApp, swap.accountToCredit.currency.name);