+++ apps/ledger-live-mobile/src/analytics/useFlushInBackground.tsx
@@ -1,0 +1,13 @@
+import { useEffect } from "react";
+import useIsAppInBackground from "../components/useIsAppInBackground";
+import { flush } from "./segment";
+
+export default function useFlushInBackground() {
+  const isAppInBakcground = useIsAppInBackground();
+
+  useEffect(() => {
+    if (isAppInBakcground) {
+      flush();
+    }
+  }, [isAppInBakcground]);
+}