+++ apps/ledger-live-mobile/src/hooks/index.ts
@@ -1,4 +1,5 @@
 import useSystem from "~/hooks/useSystem";
 import useSettings from "~/hooks/useSettings";
+import { useAccountUnit, useMaybeAccountUnit } from "~/hooks/useAccountUnit";
 
-export { useSystem, useSettings };
+export { useSystem, useSettings, useAccountUnit, useMaybeAccountUnit };