+++ libs/ledgerjs/packages/hw-app-eth/tests/ERC20/ERC20-CAL-OK.unit.test.ts
@@ -1,22 +1,23 @@
 import axios from "axios";
 import path from "path";
 import fs from "fs/promises";
-import { ethers } from "ethers";
+import { BigNumber } from "@ethersproject/bignumber";
 import { openTransportReplayer, RecordStore } from "@ledgerhq/hw-transport-mocker";
+import { serialize as serializeTransaction, type Transaction } from "@ethersproject/transactions";
 import SignatureCALEth from "../fixtures/SignatureCALEth";
 import ledgerService from "../../src/services/ledger";
 import Eth from "../../src/Eth";
 
-const transaction: ethers.Transaction = {
+const transaction: Transaction = {
   to: "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
   nonce: 14,
-  gasPrice: ethers.BigNumber.from("0x06a2bb7d00"),
-  gasLimit: ethers.BigNumber.from("0x01512c"),
-  value: ethers.BigNumber.from("0x00"),
+  gasPrice: BigNumber.from("0x06a2bb7d00"),
+  gasLimit: BigNumber.from("0x01512c"),
+  value: BigNumber.from("0x00"),
   data: "0xa9059cbb00000000000000000000000082ec3523f8a722694ca217ebfd95efbcadad77ee000000000000000000000000000000000000000000000002b5e3af16b1880000",
   chainId: 1,
 };
-const txHex = ethers.utils.serializeTransaction(transaction).slice(2);
+const txHex = serializeTransaction(transaction).slice(2);
 
 describe("ERC20 dynamic cal", () => {
   describe("ERC20 is in local CAL", () => {