+++ libs/coin-framework/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/coin-framework
 
+## 0.18.6
+
+### Patch Changes
+
+- Updated dependencies [[`65c3322`](https://github.com/LedgerHQ/ledger-live/commit/65c3322bf3871659f078148ab4b5c12b0fd53dc1)]:
+  - @ledgerhq/types-live@6.52.4
+
+## 0.18.6-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`65c3322`](https://github.com/LedgerHQ/ledger-live/commit/65c3322bf3871659f078148ab4b5c12b0fd53dc1)]:
+  - @ledgerhq/types-live@6.52.4-next.0
+
 ## 0.18.5
 
 ### Patch Changes